Struct mosquitto::Client

pub struct Client<'b, 'c, 'd> {
    pub id: String,
    pub user_name: Option<&'b str>,
    pub password: Option<&'c str>,
    pub host: Option<&'d str>,
    pub keep_alive: i32,
    pub clean_session: bool,
    pub icallbacks: HashMap<String, Box<Fn(i32)>>,
    pub scallbacks: HashMap<String, Box<Fn(&str)>>,
    pub mosquitto: *mut Struct_mosquitto,
}

Fields

id

user_name

password

host

keep_alive

clean_session

icallbacks

scallbacks

mosquitto

Methods

impl<'b, 'c, 'd> Client<'b, 'c, 'd>

fn new(id: &str, clean: bool) -> Result<Client<'b, 'c, 'd>, i32>

Creates a new mosquitto mqtt client

id: ID of the new client
clean: Clean session or not. If not, broker will remember this client(useful during connection drops)

let mut client = Client::new(&id, true).unwrap()

fn keep_alive(self, keepalive: i32) -> Self

The number of seconds after which the broker should send a PING message to the client if no other messages have been exchanged in that time. This is necessary for keeping the connection alive

let mut client = Client::new(&id, true)
                         .unwrap()
                         .keep_alive(5)

fn will(self, topic: &str, message: &str) -> Self

Will topic and message on behalf of the client. Broker will take the responsibility of publishing this after the client dies

let mut client = Client::new(&id, true)
                         .unwrap()
                         .keep_alive(5)
                         .will("goodbye", "my last words");

fn connect(&mut self, host: &'d str) -> Result<&Self, i32>

Connects the client to broker. Connects to port 1883 by default (TODO) Speciy in HOST:PORT format if you want to connect to a different port.

 match client.connect("localhost") {
     Ok(_) => println!("Connection successful --> {:?}", client.id),
     Err(n) => panic!("Connection error = {:?}", n),
 }

fn secure_connect(&mut self, host: &'d str, ca_cert: &str, client_cert: Option<(&str, &str)>) -> Result<&Self, i32>

Connects the client to broker using certificate based TLS authentication. Connects to port 8884 by default (TODO). Speciy in HOST:PORT format if you want to connect to a different port.

 match client.connect("localhost") {
     Ok(_) => println!("Connection successful --> {:?}", client.id),
     Err(n) => panic!("Connection error = {:?}", n),
 }

fn onconnect_callback<F>(&mut self, callback: F) where F: Fn(i32), F: 'static

Registered callback is called when the broker sends a CONNACK message in response to a connection. Will be called even incase of failure. All your sub/pub stuff should ideally be done in this callback when connection is successful Callback argument specifies the connection state ```ignore let i = 100;

client.onconnect_callback(move |a: i32| { println!("i = {:?}", i); println!("@@@ On connect callback {}@@@", a) }); ```

fn subscribe(&self, topic: &str, qos: Qos)

Subscibe to a topic with a Qos

 client.subscribe("hello/world", Qos::AtMostOnce);

fn onsubscribe_callback<F>(&mut self, callback: F) where F: Fn(i32), F: 'static

Registered callback will be called when broker responds to a subscription

 client.onsubscribe_callback(move |mid| {
            println!("@@@ Subscribe request received for message mid = {:?}", mid)
        });

fn publish(&self, topic: &str, message: &str, qos: Qos)

Publish a message with a Qos

 let message = format!("{}...{:?} - Message {}", count, client.id, i);
 client.publish("hello/world", &message, Qos::AtLeastOnce);

fn onpublish_callback<F>(&mut self, callback: F) where F: Fn(i32), F: 'static

Registered callback is called when a message initiated with publish has been sent to the broker successfully.

client.onpublish_callback(move |mid| {
         println!("@@@ Publish request received for message mid = {:?}", mid)
     });

fn onmesssage_callback<F>(&mut self, callback: F) where F: Fn(&str), F: 'static

Registered callback will be called when a message is received from the broker

client.onmesssage_callback(move |s| {
         println!("@@@ Message = {:?}, Count = {:?}", s, count);
     });

fn loop_forever(&self)